% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Age_Computation.R
\name{Age_Computation}
\alias{Age_Computation}
\title{Bayesian analysis for the OSL age estimation of one sample}
\usage{
Age_Computation(DATA, samplename, SavePdf = FALSE,
  OutputFileName = c("MCMCplot"), OutputFilePath = c(""),
  SaveEstimates = FALSE, OutputTableName = c("DATA"),
  OutputTablePath = c(""), BinPerSample = c(1), PriorAge = c(0.01, 100),
  LIN_fit = TRUE, Origin_fit = FALSE, distribution = c("cauchy"), I = 1,
  Taille = 50000, t = 5, Nb_chaines = 3)
}
\arguments{
\item{DATA}{list of objects: LT, sLT, ITimes, dLab, ddot_env, regDose, J, K, Nb_measurement,
provided by the function \code{\link{Generate_DataFile}}.
\code{DATA} can contain information for more than one sample.}

\item{samplename}{character: name of the sample.}

\item{SavePdf}{boolean (with default): if TRUE save graph in pdf file named \code{OutputFileName} in folder \code{OutputFilePath}.}

\item{OutputFileName}{character (with default): name of the pdf file that will be generated by the function if \code{SavePdf}=TRUE.}

\item{OutputFilePath}{character (with default): path to the pdf file that will be generated by the function if \code{SavePdf}=TRUE.}

\item{SaveEstimates}{boolean (with default): if TRUE save Bayes estimates and confidence interval at level 68\% and 95\%,
in a csv table named \code{OutputFileName} in folder \code{OutputFilePath}.}

\item{OutputTableName}{character (with default): name of the table that will be generated by the function if \code{SaveEstimates}=TRUE.}

\item{OutputTablePath}{character (with default): path to the table that will be generated by the function if \code{SaveEstimates}=TRUE.}

\item{BinPerSample}{integer vector (with default): vector with the number of BIN files per sample.
If in \code{DATA} there is more than one sample,
the \code{BinPerSample} vector must be the same as that used to run the function
\code{\link{Generate_DataFile}} for generating the \code{DATA} object.}

\item{PriorAge}{numeric vector (with default): lower and upper bounds for the sample age parameter.
\code{length(PriorAge)=2}.}

\item{LIN_fit}{logical (with default): if TRUE (default) allows a linear component,
on top of the (default) saturating exponential curve, for the fitting of dose response curves.
Please see details for more informations on the proposed dose response curves.}

\item{Origin_fit}{logical (with default): if TRUE, forces the dose response curves to pass through the origin.
Please see details for more informations on the proposed growth curves.}

\item{distribution}{character (with default): type of distribution that defines
how individual equivalent dose values are distributed around the palaeodose.
Allowed inputs are \bold{"cauchy"}, \bold{"gaussian"} and \bold{"lognormal"}.}

\item{I}{integer (with default): if \code{DATA} contains data from more than one sample,
I indicates the ID number of the sample to be analysed.}

\item{Taille}{integer (with default): number of iterations for the MCMC computation (for more information see \code{\link{jags.model}}).}

\item{t}{integer (with default): 1 every \code{t} iterations of the MCMC is considered for sampling the posterior distribution
(for more information see \code{\link{jags.model}}).}

\item{Nb_chaines}{integer (with default): number of independent chains for the model (for more information see \code{\link{jags.model}}).}
}
\value{
\bold{1- NUMERICAL OUTPUT}\cr

\bold{A list containing the following objects:}
\itemize{
 \item \bold{Sampling} that corresponds to a sample of the posterior distributions
 of the Age, palaeodose and equivalent dose dispersion parameters.
 \item \bold{Model_GrowthCurve}, stating which dose response fitting option was chosen;
 \item \bold{Distribution}, stating which distribution was chosen to model the dispersion of
 individual equivalent dose values around the palaeodose of the sample;
 \item \bold{PriorAge}, stating the priors used for the age parameter.
}

\bold{The Gelman and Rudin test of convergency}

Prints the result of the Gelman and Rudin test of convergency for the age, palaeodose and equivalent dose dispersion parameters.
A result close to one is expected.\cr
In addition, the user must visually assess the convergency of the trajectories by looking at the pdf file
generated by the function (see 2- for more informations).\cr
If both convergencies (Gelman and Rudin test and plot checking) are satisfactory,
the user can consider the printed estimates as valid. Otherwise, the user may try increasing the number of MCMC interations
(\code{Taille}) to reach convergency.

\bold{Credible intervals and Bayes estimates}

Prints the Bayes esitmates, the credible intervals at level 95\% and 68\% for
the age, palaeodose and equivalent dose dispersion parameters of the sample.

\bold{2- PLOT OUTPUT}

A pdf file with the MCMC trajectories and posterior distributions of the age, palaeodose and equivalent dose dispersion parameters.\cr
The first line of the figure correponds to the age parameter, the second to the palaeodose parameter and the third to the equivalent dose dispersion parameter.
On each line, the plot on the left represents the MCMC trajectories, and the one on the right the posterior distribution of the parameter.
}
\description{
This function computes the age of a sample according to the model developed in Combes and Philippe (2017),
based on an output of the \code{\link{Generate_DataFile}} function.\cr
A sample, for which data is avalilable in several BIN files, can be analysed.
}
\details{
For more flexibility, the user can choose from 4 dose response curves:
\itemize{
  \item \bold{Saturating exponential plus linear growth} (\code{AgeMultiBF_EXPLIN}):

  for all \code{x} in IR+, \code{f(x)=a(1-exp(-x/b))+cx+d}; select
  \itemize{
    \item \code{LIN_fit=TRUE}
    \item \code{Origin_fit=FALSE}
  }
  \item \bold{Saturating exponential growth} (\code{AgeMultiBF_EXP}):

  for all \code{x} in IR+, \code{f(x)=a(1-exp(-x/b))+d}; select
  \itemize{
    \item \code{LIN_fit=FALSE}
    \item \code{Origin_fit=FALSE}
  }
  \item \bold{Saturating exponential plus linear growth and fitting through the origin} (\code{AgeMultiBF_EXPLINZO}):

  for all \code{x} in IR+, \code{f(x)=a(1-exp(-x/b))+cx}; select
  \itemize{
    \item \code{LIN_fit=TRUE}
    \item \code{Origin_fit=TRUE}
  }
  \item \bold{Saturating exponential growth and fitting through the origin} (\code{AgeMultiBF_EXPZO}):

  for all \code{x} in IR+, \code{f(x)=a(1-exp(-x/b))}; select
  \itemize{
    \item \code{LIN_fit=FALSE}
    \item \code{Origin_fit=TRUE}
  }
}
}
\examples{
## load data file generated by the function Generate_DataFile
# data(DATA1,envir = environment())
# Age=Age_Computation(DATA1,samplename="GDB3",SavePdf=FALSE,Taille=1000)

}
\references{
Combes, Benoit and Philippe, Anne, 2017.
Bayesian analysis of multiplicative Gaussian error for multiple ages estimation in optically stimulated luminescence dating.
Quaternary Geochronology (in press)

Combes, B., Philippe, A., Lanos, P., Mercier, N., Tribolo, C., Guerin, G., Guibert, P., Lahaye, C., 2015.
A Bayesian central equivalent dose model for optically stimulated luminescence dating.
Quaternary Geochronology 28, 62-70. doi:10.1016/j.quageo.2015.04.001
}
\seealso{
\code{\link{Generate_DataFile}}, \code{\link{rjags}} package, \code{\link{MCMC_plot}}
}
\author{
Claire Christophe, Guillaume Guerin
}
