% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgeC14_Computation.R
\name{AgeC14_Computation}
\alias{AgeC14_Computation}
\title{Bayesian analysis for age estimation of 14C Cal age of various samples}
\usage{
AgeC14_Computation(Data_C14Cal, Data_SigmaC14Cal, Nb_sample, SampleNames,
  SavePdf = FALSE, OutputFileName = c("MCMCplot", "HPD_Cal14CCurve",
  "summary", "PosteriorDensity", "2per2Laws"), OutputFilePath = c(""),
  SaveEstimates = FALSE, OutputTableName = c("DATA"),
  OutputTablePath = c(""), PriorAge = rep(c(10000, 50000), Nb_sample),
  StratiConstraints = c(), sepSC = c(","), Model = c("full"),
  CalibrationCurve = c("AtmosphericNorth"), Taille = 50000, t = 5,
  Nb_chaines = 3)
}
\arguments{
\item{Data_C14Cal}{numeric vector: corresponding to 14C Cal age estimate.}

\item{Data_SigmaC14Cal}{numeric vector: correponding to the error of 14C Cal age estimates.}

\item{Nb_sample}{integer: number of samples, \code{Nb_sampe>1}.}

\item{SampleNames}{character vector: names of sample. The length of this vector is equal to \code{Nb_sample}.}

\item{SavePdf}{boolean (with default): if TRUE save graphs in pdf file named \code{OutputFileName} in folder \code{OutputFilePath}.}

\item{OutputFileName}{character (with default): name of the pdf file that will be generated by the function if \code{SavePdf}=TRUE.}

\item{OutputFilePath}{character (with default): path to the pdf file that will be generated by the function if \code{SavePdf}=TRUE.}

\item{SaveEstimates}{boolean (with default): if TRUE save Bayes estimates, credible interval at level 68\% and 95\% and ,
in a csv table named \code{OutputFileName} in folder \code{OutputFilePath}.}

\item{OutputTableName}{character (with default): name of the table that will be generated by the function if \code{SaveEstimates}=TRUE.}

\item{OutputTablePath}{character (with default): path to the table that will be generated by the function if \code{SaveEstimates}=TRUE.}

\item{PriorAge}{numeric vector (with default): lower and upper bounds for age parameter of each sample.
\code{length(PriorAge)=2*Nb_sample}
and \code{PriorAge[2i-1,2i]} correponds to the lower and upper bounds of sample whose number ID is equal to \code{i}.}

\item{StratiConstraints}{numeric matrix or character(with default): input object for the statigraphic relation between samples.
If there is stratigraphic relation between samples see the details section for instructions regarding how to correctly fill \code{StratiConstraints}.
Otherwise, default value is suitable.}

\item{sepSC}{character (with default): if \code{StratiConstraints} is character,
indicate column separator in \code{StratiConstraints} .csv file.}

\item{Model}{character (with default): if \bold{"full"}, error on estimate calibration curve is taken account.
If \bold{"naive"} this error is not taken account in the age estimate.}

\item{CalibrationCurve}{character (with default): calibration curve choosen.Allowed inputs are
\itemize{
  \item \bold{"AtmosphericNorth"} for Northern Hemisphere atmospheric radiocarbon calibration curve,
  \item \bold{"Marine"} for Marine radiocarbon calibration curve,
  \item \bold{"AtmosphericSouth"} for Southern Hemisphere atmospheric radiocarbon calibration curve
}}

\item{Taille}{integer (with default): number of iterations for the MCMC computation (for more information see \code{\link{jags.model}}).}

\item{t}{integer (with default): 1 every \code{t} iterations of the MCMC is considered for sampling the posterior distribution
(for more information see \code{\link{jags.model}}).}

\item{Nb_chaines}{integer (with default): number of independent chains for the model (for more information see \code{\link{jags.model}}).}
}
\value{
\bold{NUMERICAL OUTPUT}\cr

\enumerate{
\item \bold{A list containing the following objects:}
 \itemize{
  \item \bold{Sampling}: that corresponds to a sample of the posterior distributions of the age parameters.
  \item \bold{Model}, stating which model was chosen (\code{"full"} or \code{"naive"});
  \item \bold{CalibrationCurve}, stating which radiocarbon calibration curve was chosen;
  \item \bold{PriorAge}, stating the priors used for the age parameter;
  \item \bold{StratiConstraints}, stating the stratigraphic relations between samples considered in the model.
 }

  \item\bold{The Gelman and Rudin test of convergency}: prints the result of the Gelman and Rudin test of convergency for the age estimate for each sample.
A result close to one is expected.\cr
In addition, the user must visually assess the convergency of the trajectories by looking at the graph
generated by the function (see \bold{PLOT OUTPUT} for more informations).\cr
If both convergencies (Gelman and Rudin test and plot checking) are satisfactory,
the user can consider the estimates as valid.
Otherwise, the user may try increasing the number of MCMC interations (\code{Taille})
or be more precise on the \code{PriorAge} parameter to reach convergency.
  \item \bold{Credible intervals and Bayes estimates}: prints the Bayes esitmates, the credible intervals at 95\% and 68\% for
the age parameters for each sample.
}

\bold{PLOT OUTPUT}

\enumerate{
 \item\bold{MCMC trajectories}: A graph with the MCMC trajectories and posterior distributions of the age parameter is displayed. \cr
On each line, the plot on the left represents the MCMC trajectories, and the one on the right the posterior distribution of the parameter.
 \item \bold{Age estimate and HPD on calibration curve}: plot age estimate and HPD on calibration plot to check...........
 \item \bold{Summary of sample age estimates}: plot credible intervals and Bayes estimate of each sample age on a same graph.
 \item \bold{Posterior density}: plot posterior density of each age estimate.
 \item\bold{Matrix of scatterplots}: the \code{i},\code{j} th scatterplot contains sample of the posterior distribution of the sample ID \code{i} plotted
 against the one of the sample ID \code{j}.
}
}
\description{
This function calibrate 14C Cal age of samples to get an age (years).
The user can choose one of the following radiocarbon calibration curve:
Northern or Sourthen Hemisphere or marine atmospheric.
}
\details{
\bold{** How to fill} \code{StratiConstraints} \bold{? **}

If there is stratigraphic relations between samples, \bold{14C Cal estimate age in Data_C14Cal must be ordered by order of increasing ages}.

The user can fill the \code{StratiConstraints} matrix as follow.\cr
\enumerate{
 \item \bold{Size of the matrix}: row number of \code{StratiConstraints} matrix is equal to \code{Nb_sample+1},
and column number is equal to \code{Nb_sample}. \cr
 \item \bold{First line of the matrix}:
for all \code{i in {1,...,Nb_Sample}}, \code{StratiConstraints[1,i]=1} that means the lower bound of the sample age (given in \code{PriorAge[2i-1]})
for the sample whose number ID is equal to \code{i}, is taken into account.\cr
 \item \bold{Sample relations}: for all  \code{j in {2,...,Nb_Sample+1}} and all \code{i in {j,...,Nb_Sample}},
\code{StratiConstraints[j,i]=1} if sample age whose number ID is equal to \code{j-1} is lower than sample age whose number ID is equal to \code{i}.
Otherwise, \code{StratiConstraints[j,i]=0}.
}
Note that \code{StratiConstraints_{2:Nb_sample+1,1:Nb_sample}} is a upper triangular matrix.

The user can also use \code{\link{SCMatrix}} function to construc the \code{StratiConstraints} matrix.

The user can also refer to a .csv file that containts the relation between samples as defined above.
}
\examples{
## Load data
## Age computation of samples...
# Age=AgeC14_Computation(Data_C14Cal=Data_C14Cal,Data_SigmaC14Cal=Data_SigmaC14Cal,
#     Nb_sample=Nb_sample,SampleNames=SampleNames,PriorAge=PriorAge,Taille=1000)

}
\references{
Reimer PJ, Bard E, Bayliss A, Beck JW, Blackwell PC, Bronl Ramsey C, Buck CE, Cheng H, Edwards RL, Friedrich M,
Grootes PM, Guilderson TP, Haflidason H, Hajdas I, Hatte C, Heaton TJ, Hoffmann DL, Hogg AG, Hughen KA, Kaiser KF, Kromer B,
Manning SW, Niu M, Reimer RW, Richards DA, Scott EM, Southon JR, Staff RA, Turney CSM, van der Plicht J. 2013.
IntCal13 ans Marine13 radiocarbon age calibration curves 0-50000 years cal BP. Radiocarbon 55(4)=1869-1887.

Hogg AG, Hua Q, Blackwell PG, Niu M, Buck CE, Guilderson TP, Heaton TJ, Palmer JG, Reimer PJ, Reimer RW, Turney CSM, Zimmerman SRH.
2013. SHCal13 Southern Hemisphere calibration, 0-50000 years cal BP. Radiocarbon 55(4):1889-1903
}
\seealso{
\code{\link{rjags}}, \code{\link{MCMC_plot}}, \code{\link{SCMatrix}}
}
\author{
Claire Christophe, Guillaume Guerin
}
