% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgeS_Computation.R
\name{AgeS_Computation}
\alias{AgeS_Computation}
\title{Bayesian analysis for the OSL age estimation of various samples}
\usage{
AgeS_Computation(DATA, Nb_sample, BinPerSample = rep(1, Nb_sample),
  SampleNames, SavePdf = FALSE, OutputFileName = c("MCMCplot", "summary",
  "2per2Laws"), OutputFilePath = c(""), SaveEstimates = FALSE,
  OutputTableName = c("DATA"), OutputTablePath = c(""), THETA = c(),
  sepTHETA = c(","), PriorAge = rep(c(0.01, 100), Nb_sample),
  StratiConstraints = c(), sepSC = c(","), LIN_fit = TRUE,
  Origin_fit = FALSE, distribution = c("cauchy"), Taille = 50000, t = 5,
  Nb_chaines = 3)
}
\arguments{
\item{DATA}{list of objects: LT, sLT, ITimes, dLab, ddot_env, regDose, J, K, Nb_measurement, 
provided by the function \code{\link{Generate_DataFile}}. 
\code{DATA} contains information for more than one sample.}

\item{Nb_sample}{integer \bold{(required)}: number of samples, \code{Nb_sampe>1}.}

\item{BinPerSample}{integer vector (with default): vector with the number of BIN files per sample. 
The length of this vector is equal to \code{Nb_sample}. 
\code{BinPerSample}[i] correponds to the number of BIN files for the sample whose number ID is equal to \code{i}.
For more information to fill this vector, we refer to detatils in \code{\link{Generate_DataFile}}.}

\item{SampleNames}{character vector: names of sample. The length of this vector is equal to \code{Nb_sample}.}

\item{SavePdf}{boolean (with default): if TRUE save graph in pdf file named \code{OutputFileName} in folder \code{OutputFilePath}.}

\item{OutputFileName}{character (with default): name of the pdf file that will be generated by the function if \code{SavePdf}=TRUE.}

\item{OutputFilePath}{character (with default): path to the pdf file that will be generated by the function if \code{SavePdf}=TRUE.}

\item{SaveEstimates}{boolean (with default): if TRUE save Bayes estimates and confidence interval at level 68\% and 95\%,
in a csv table named \code{OutputFileName} in folder \code{OutputFilePath}.}

\item{OutputTableName}{character (with default): name of the table that will be generated by the function if \code{SaveEstimates}=TRUE.}

\item{OutputTablePath}{character (with default): path to the table that will be generated by the function if \code{SaveEstimates}=TRUE.}

\item{THETA}{numeric matrix or character (with default): input object for systematic and individual error.
If systematic errors are considered, see the details section for instructions regarding how to correctly fill \code{THETA}.
Otherwise, default value is suitable, and only individual error is considered.}

\item{sepTHETA}{character (with default): if \code{THETA} is character, 
indicate column separator in \code{THETA} .csv file.}

\item{PriorAge}{numeric vector (with default): lower and upper bounds for age parameter of each sample.
\code{length(PriorAge)=2*Nb_sample}
and \code{PriorAge[2i-1,2i]} correponds to the lower and upper bounds of sample whose number ID is equal to \code{i}.}

\item{StratiConstraints}{numeric matrix or character(with default): input object for the statigraphic relation between samples.
If there is stratigraphic relation between samples see the details section for instructions regarding how to correctly fill \code{StratiConstraints}.
Otherwise, default value is suitable.}

\item{sepSC}{character (with default): if \code{StratiConstraints} is character, 
indicate column separator in \code{StratiConstraints} .csv file.}

\item{LIN_fit}{logical (with default): if TRUE (default) allows a linear component, 
on top of the (default) saturating exponential curve, for the fitting of dose response curves.
Please see details for more informations on the proposed dose response curves.}

\item{Origin_fit}{logical (with default): if TRUE, forces the dose response curves to pass through the origin.
Please see details for more informations on the proposed growth curves.}

\item{distribution}{character (with default): type of distribution that defines
how individual equivalent dose values are distributed around the palaeodose. 
Allowed inputs are \bold{"cauchy"}, \bold{"gaussian"} and \bold{"lognormal"}.}

\item{Taille}{integer (with default): number of iterations for the MCMC computation (for more information see \code{\link{jags.model}}).}

\item{t}{integer (with default): 1 every \code{t} iterations of the MCMC is considered for sampling the posterior distribution
(for more information see \code{\link{jags.model}}).}

\item{Nb_chaines}{integer (with default): number of independent chains for the model (for more information see \code{\link{jags.model}}).}
}
\value{
\bold{1- NUMERICAL OUTPUT}\cr

\bold{A list containing the following objects:}
\itemize{
 \item \bold{Sampling} that corresponds to a sample of the posterior distributions 
 of the Age, palaeodose and equivalent dose dispersion parameters.
 \item \bold{Model_GrowthCurve}, stating which dose response fitting option was chosen;
 \item \bold{Distribution}, stating which distribution was chosen to model the dispersion of 
 individual equivalent dose values around the palaeodose of the sample; 
 \item \bold{PriorAge}, stating the priors used for the age parameter. 
}

\bold{The Gelman and Rudin test of convergency}

Prints the result of the Gelman and Rudin test of convergency for the age, palaeodose and equivalent dose dispersion parameters for each sample. 
A result close to one is expected.\cr
In addition, the user must visually assess the convergency of the trajectories by looking at the pdf file  
generated by the function (see 2- for more informations).\cr
If both convergencies (Gelman and Rudin test and plot checking) are satisfactory, 
the user can consider the printed estimates as valid. Otherwise, the user may try increasing the number of MCMC interations 
(\code{Taille}) to reach convergency.

\bold{Credible intervals and Bayes estimates}

Prints the Bayes esitmates, the credible intervals at level 95\% and 68\% for 
the age, palaeodose and equivalent dose dispersion parameters for each sample.

\bold{2- PLOT OUTPUT}

\bold{MCMC trajectories}

A pdf file with the MCMC trajectories and posterior distributions of the age, palaeodose and equivalent dose dispersion parameters.
There is one page per sample. \cr
The first line of the figure correponds to the age parameter, the second to the palaeodose parameter and the third to the equivalent dose dispersion parameter. 
On each line, the plot on the left represents the MCMC trajectories, and the one on the right the posterior distribution of the parameter. 

\bold{Summary of sample ages}

Plot credible intervals and bayes estimate of each sample age on a same graph.

\bold{Stratigraphic constraints}

If there are stratigraphic constraints, plot sampling of each sample two by two...
}
\description{
This function computes the age of at leat two samples according to the model developed in Combes and Philippe (2017), 
based on an output of the \code{\link{Generate_DataFile}} function.\cr 
Samples, for which data is avalilable in several BIN files, can be analysed.
}
\details{
\bold{** How to fill} \code{StratiConstraints} \bold{? **} 

If there is stratigraphic relations between samples, information in DATA must be ordered by order of increasing ages. 
The user can fill the \code{StratiConstraints} matrix as follow.\cr
First, concerning the \bold{size of the matrix}, row number of \code{StratiConstraints} matrix is equal to \code{Nb_sample+1}, 
and column number is equal to \code{Nb_sample}. \cr
Secondly, conserning the \bold{first line of the matrix}, 
for all \code{i in {1,...,Nb_Sample}}, \code{StratiConstraints[1,i]=1} that means the lower bound of the sample age given in \code{PriorAge[2i-1]} 
for the sample whose number ID is equal to \code{i}, is taken into account.\cr
Thirdly, concerning the \bold{sample relations}, for all  \code{j in {2,...,Nb_Sample+1}} and all \code{i in {j,...,Nb_Sample}},
\code{StratiConstraints[j,i]=1} if sample age whose number ID is equal to \code{j-1} is lower than sample age whose number ID is equal to \code{i}.
Otherwise, \code{StratiConstraints[j,i]=0}.\cr
Note that \code{StratiConstraints_{2:Nb_sample+A,1:Nb_sample}} is a upper triangular matrix.

The user can also use \code{\link{SCMatrix}} function to construc the \code{StratiConstraints} matrix.

The user can also refer to a .csv file that containts the relation between samples as defined above.  

\bold{** How to fill} \code{THETA} \bold{ concerning common and individual error? **} 

If systematic errors are considered, the user can fill the \code{THETA} matrix as follow.\cr
Note that row number of \code{THETA} is equal the column number. 
For all \code{i in {1,...,Nb_sample}}, \code{THETA[i,i]} containts individual error plus systematic error of the sample whose number ID is equal to \code{i}.
For all \code{i,j in {1,...,Nb_sample}} and \code{i} different from \code{j} , 
\code{THETA[i,j]} containts common error between samples whose number ID are equal to \code{i} and \code{j}.\cr
Note that \code{THETA[i,j]} is a symetric matrix.

The user can also refer to a .csv file that containts the errors as defined above.

\bold{** Different growth curves **}

As for \code{Age_Computation}, the user can choose from 4 dose response curves:
\itemize{
  \item \bold{Saturating exponential plus linear growth} (\code{AgeMultiBF_EXPLIN}):
  
  for all \code{x} in IR+, \code{f(x)=a(1-exp(-x/b))+cx+d}; select
  \itemize{
    \item \code{LIN_fit=TRUE}
    \item \code{Origin_fit=FALSE}
  }
  \item \bold{Saturating exponential growth} (\code{AgeMultiBF_EXP}):
  
  for all \code{x} in IR+, \code{f(x)=a(1-exp(-x/b))+d}; select
  \itemize{
    \item \code{LIN_fit=FALSE}
    \item \code{Origin_fit=FALSE}
  }
  \item \bold{Saturating exponential plus linear growth and fitting through the origin} (\code{AgeMultiBF_EXPLINZO}):
  
  for all \code{x} in IR+, \code{f(x)=a(1-exp(-x/b))+cx}; select
  \itemize{
    \item \code{LIN_fit=TRUE}
    \item \code{Origin_fit=TRUE}
  }
  \item \bold{Saturating exponential growth and fitting through the origin} (\code{AgeMultiBF_EXPZO}):
  
  for all \code{x} in IR+, \code{f(x)=a(1-exp(-x/b))}; select
  \itemize{
    \item \code{LIN_fit=FALSE}
    \item \code{Origin_fit=TRUE}
  }
}
}
\examples{
## Load data
# data(DATA1,envir = environment())
# data(DATA2,envir = environment()) 
# DATA=Concat_DataFile(DATA2,DATA1)

## Age computation of samples GDB5 and GDB3, 
## without common error and without stratigraphic constraints
# Age=AgeS_Computation(DATA,Nb_sample=2,SampleNames=c("GDB5","GDB3"),SavePdf=FALSE,Taille=1000)
 
## Age computation of samples GDB5 and GDB3, 
## without common error, assuming GDB5 age younder than GDB3 age
# SC=matrix(data=c(1,1,0,1,0,0),ncol=Nb_sample,nrow = (Nb_sample+1),byrow = T)
# Age=AgeS_Computation(DATA,Nb_sample=2,SampleNames=c("GDB5","GDB3"),
#    StratiConstraints=SC,Taille=10000)

}
\references{
Combes, Benoit and Philippe, Anne, 2017.
Bayesian analysis of multiplicative Gaussian error for multiple ages estimation in optically stimulated luminescence dating.
Quaternary Geochronology (in press)

Combes, B., Philippe, A., Lanos, P., Mercier, N., Tribolo, C., Guerin, G., Guibert, P., Lahaye, C., 2015. 
A Bayesian central equivalent dose model for optically stimulated luminescence dating. 
Quaternary Geochronology 28, 62-70. doi:10.1016/j.quageo.2015.04.001
}
\seealso{
\code{\link{Generate_DataFile}}, \code{\link{rjags}} package, \code{\link{MCMC_plot}}, 
\code{\link{Age_Computation}}, \code{\link{SCMatrix}}
}
\author{
Claire Christophe, Guillaume Guerin
}
