% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayLum-package.R
\docType{package}
\name{BayLum-package}
\alias{BayLum-package}
\alias{BayLum}
\title{Chronological Bayesian Models Integrating Optically Stimulated Luminescence Dating}
\description{
Chronological Bayesian Models Integrating Optically Stimulated Luminescence Dating
}
\keyword{package}
