% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Concat_DataFile.R
\name{Concat_DataFile}
\alias{Concat_DataFile}
\title{Concatenates two outputs of the function \code{\link{Generate_DataFile}}}
\usage{
Concat_DataFile(u1, u2)
}
\arguments{
\item{u1}{list of objects: LT, sLT, ITimes, dLab, ddot_env, regDose, J, K, Nb_measurement.}

\item{u2}{list of objects: LT, sLT, ITimes, dLab, ddot_env, regDose, J, K, Nb_measurement.}
}
\value{
A List of objects: LT, sLT, ITimes, dLab, ddot_env, regDose, J, K, Nb_measurement.
}
\description{
This function allows concatenating two lists provided as output of the \code{\link{Generate_DataFile}} function.\cr 
Only concatenation of 2 files is possible.
}
\examples{
# load data files
data(DATA1,envir = environment())
data(DATA2,envir = environment())
# concatenate two data files 
DATA3=Concat_DataFile(DATA1,DATA2)
str(DATA3)
}
\seealso{
\code{\link{Generate_DataFile}}
}
\author{
Claire Christophe, Guillaume Guerin
}
