% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LT_RegenDose.R
\name{LT_RegenDose}
\alias{LT_RegenDose}
\title{Plots L/T as a function of regenerative dose}
\usage{
LT_RegenDose(DATA, Path, Names, BinPerSample = rep(1), SampleNumber = 1,
  sepDP = c(","))
}
\arguments{
\item{DATA}{list of objects: LT, sLT, ITimes, dLab, ddot_env, regDose, J, K, Nb_measurement, provided by the function \code{\link{Generate_DataFile}}.
\code{DATA} can contain information from more than one sample.}

\item{Path}{character: path to the project folder (the same as the one used in \code{\link{Generate_DataFile}} to provide \code{DATA})}

\item{Names}{character vector:  list of names of the sub-folders containing the BIN files, 
which were used by \code{\link{Generate_DataFile}} to generate the \code{DATA} object.}

\item{BinPerSample}{integer vector (with default): vector with the number of BIN files per sample,
which was used in \code{\link{Generate_DataFile}} to generate the \code{DATA} object.}

\item{SampleNumber}{integer(with default): ID number (in [1,\code{Nb_sample}]) of the sample selected for plotting L/T as a function of regenerative doses. 
Required if the \code{DATA} object contains information for more than one sample.}

\item{sepDP}{character(with default): column separator in the DiscPose.csv file.}
}
\value{
L/T plots; there are as many plots as selected aliquots in the DiscPos.csv file. There are 9 plots per page. 
There is not interpolation.
}
\description{
This function plots L/T values as a function of Regenerative Dose, for every selected aliquot.
}
\details{
To fill \code{Names} and \code{BinPerSample}, we refer to the detail section from the \code{\link{Generate_DataFile}} function.
}
\examples{
## load data file generated by the function Generate_DataFile
# data(DATA1,envir = environment()) 
# LT_RegenDose(DATA1,Path=c('inst/extdata/'),Names=c("GDB3"))

}
\seealso{
\code{\link{Generate_DataFile}}
}
\author{
Claire Christophe, Guillaume Guerin
}
