% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Ages.R
\name{plot_Ages}
\alias{plot_Ages}
\title{Create age plot}
\usage{
plot_Ages(object, ...)
}
\arguments{
\item{object}{\link{list} (\strong{required}): Output as created by functions like \link{AgeC14_Computation}.}

\item{...}{further arguments to control the plot output}
}
\value{
A plot
}
\description{
This function creates an age plot showing the mean ages and along with the credible intervals.
}
\examples{
##TODO

}
\seealso{
\link{AgeC14_Computation}
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, CNRS - Université Bordeaux Montaigne (France), based on code
written by Claire Christophe
}
