% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MCMC.R
\name{plot_MCMC}
\alias{plot_MCMC}
\title{Plot MCMC trajectories and posterior distributions}
\usage{
plot_MCMC(object, axes_labels = c(A = "Age (ka)", D = "D (Gy)", sD =
  "sD (Gy)"), n.chains = NULL, n.iter = 1000, smooth = FALSE,
  plot_single = FALSE, ...)
}
\arguments{
\item{object}{\link[coda:mcmc.list]{coda::mcmc.list} or \link[coda:mcmc]{coda::mcmc} (\strong{required}): Output generated by \link[rjags:jags.model]{rjags::jags.model},
e.g., in \code{\link{Age_Computation}}}

\item{axes_labels}{\link{character} (with default): Axes labels used for the trace and density plots. The labels should
be provided as names \link{character} \link{vector} with the parameter names as the names used to asign the axes labelling.
The labelling for the xaxis (trace plots) and yaxis (density plot) cannot be modified.}

\item{n.chains}{\link{numeric} (optional): Set the number of chains to visualise, if nothing is provided the
number of chains is determined from the input object}

\item{n.iter}{\link{numeric} (with default): Set the number of iterations to be visualised in the trace plots.}

\item{smooth}{\link{logical} (with default): Enable/disables smooth of trace plots using \link[stats:smooth]{stats::smooth}}

\item{plot_single}{\link{logical} (with default): Enables/disables the single plot mode of the function, i.e.
if set to \code{TRUE} every plot is returned in a single plot and own \link{par} settings can be applied.}

\item{...}{further arguments that can be passed to modifiy the plot output. Supported arguments are
\code{lwd}, \code{lty}, \code{col}, \code{type}, \code{cex},\code{mtext}, cf. \link{mtext} for \code{mtext} and \link{plot.default} for the other
arguments.}
}
\value{
Two plots: Traces of the MCMC chains and the corresponding density plots. This plots
are similar to \link[coda:traceplot]{coda::traceplot} and \link[coda:densplot]{coda::densplot}.
}
\description{
This function uses the output of \link[rjags:jags.model]{rjags::jags.model} to visualise the traces of the MCMC and the
corresponding densities. In particular it displays the posterior distributions of the age, if it is calculated,
palaeodose and the equivalent dose dispersion parameters of the sample. The function output is very
similar to plot output produced with the 'coda' package, but tailored to meet the needs in
the context of the 'BayLum' package.
}
\details{
The function is used in the function \code{\link{Age_Computation}}, \code{\link{AgeS_Computation}}
and \code{\link{Palaeodose_Computation}}, but can be used also as standalone plot function.
}
\section{Function version}{
 0.1.0
}

\examples{
data(MCMCsample,envir = environment())
object <- coda::as.mcmc(MCMCsample)
plot_MCMC(object)

}
\seealso{
\code{\link{Age_Computation}}, \code{\link{AgeS_Computation}}, \code{\link{Palaeodose_Computation}},
\code{\link{coda.samples}} and \code{\link{rjags}} packages.
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, Université Bordeaux Montaigne (France). This function
is a re-written version of the function \code{MCMC_plot()} by Claire Christophe
}
\keyword{dplot}
