% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Scatterplots.R
\name{plot_Scatterplots}
\alias{plot_Scatterplots}
\alias{ScatterSamples}
\title{Display Scatter Plot Matrix of the Bayesian Age Results}
\usage{
plot_Scatterplots(object, variables = c("A"), sample_names = NULL,
  sample_selection = NULL, n.chains = NULL, ...)

ScatterSamples(...)
}
\arguments{
\item{object}{\link[coda:mcmc.list]{coda::mcmc.list} (\strong{required}): mcmc list objects generated by \link[rjags:jags.model]{rjags::jags.model} in \link{AgeS_Computation},
\link{AgeC14_Computation} or \link{Age_OSLC14}.}

\item{variables}{\link{character} (\emph{with default}): variable to be selected for the scatter plot, e.g., \code{"A"}. Please note
that you can only select one variable at the time}

\item{sample_names}{\link{character} (\emph{optional}): sample names shown in the plot matrix}

\item{sample_selection}{\link{numeric} (\emph{with default}): vector of samples to be plotted in the scatter matrix, e.g.,
\code{c(1,2)} will plot the first two samples, \code{c(1,3)} will plot samples 1 and 3 and \code{c(1:3)} will plot the first
three samples}

\item{n.chains}{\link{integer} (\emph{with default}): allows to limit the number of chains shown,
by default the results of all chains are plotted.}

\item{...}{further arguments to control the plot output, supported are \code{main}, \code{xlab}, \code{ylab}, \code{colramp}, \code{pscales}}
}
\value{
A scatter plot based on \link[hexbin:hexplom]{hexbin::hexplom}
}
\description{
Create a hexbin plot matrix (\link[hexbin:hexplom]{hexbin::hexplom}) of age results returned by the bayesian age calculation.
}
\section{Function version}{
 0.1.0
}

\examples{
data(AgeS,envir = environment())
plot_Scatterplots(
   object = AgeS$Sampling,
   sample_names = c("GDB5", "GDB3"),
   sample_selection = c(1,2)
 )

}
\seealso{
\link{Age_Computation}, \link{AgeS_Computation}, \link{AgeC14_Computation},
and \link{rjags} packages.
}
\author{
Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, CNRS - Université Bordeaux Montaigne (France), based on the
function \code{ScatterSamples()} by Claire Christophe, Anne Philippe, Guillaume Guérin
}
