% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SC_Ordered.R
\name{SC_Ordered}
\alias{SC_Ordered}
\title{Stratigraphic matrix for samples ordered}
\usage{
SC_Ordered(Nb_sample)
}
\arguments{
\item{Nb_sample}{interger: the sample number.}
}
\value{
Stratigraphic matrix where each sample are ordered by incresing order
This matrix can be intergrate in \code{AgeS_Computation} function.
We refer to detail on \code{AgeS_Computation} for more information concerning this matrix.
}
\description{
Construc the stratigraphic matrix used in fonction \code{AgeS_Computation} and \code{AgeC14_Computation}
for samples that are all ordered by increasing oder.
}
\examples{
# compute the stratigraphic matrix for 3 samples such that the first sample is younger
# than the second, and the second is younger than the third
(SC=SC_Ordered(Nb_sample=3))

}
\seealso{
\code{AgeS_Computation}, \code{SC_Matrix}
}
\author{
Claire Christophe, Anne Philippe, Guillaume Guerin
}
